Module popupRegistrationValidation.js was created
